+++ ietfYangDraftPull/pull_local.py
@@ -38,7 +38,7 @@ from utility.create_config import create_config
 from utility.script_config_dict import script_config_dict
 from utility.scriptConfig import ScriptConfig
 from utility.staticVariables import github_url
-from utility.util import job_log
+from utility.util import JobLogMessage, job_log
 
 BASENAME = os.path.basename(__file__)
 FILENAME = BASENAME.split('.py')[0]
@@ -66,6 +66,7 @@ def run_populate_script(directory: str, notify: bool, logger: logging.Logger) ->
         script_conf.args.__setattr__('sdo', True)
         script_conf.args.__setattr__('dir', directory)
         script_conf.args.__setattr__('notify_indexing', notify)
+        script_conf.args.__setattr__('official_source', 'ietf')
         logger.info(f'Running populate.py script over {directory}')
         submodule.main(script_conf=script_conf)
     except Exception:
@@ -101,7 +102,7 @@ def populate_directory(directory: str, notify_indexing: bool, logger: logging.Lo
 
 
 @job_log(file_basename=BASENAME)
-def main(script_conf: ScriptConfig = DEFAULT_SCRIPT_CONFIG.copy()) -> list[dict[str, str]]:
+def main(script_conf: ScriptConfig = DEFAULT_SCRIPT_CONFIG.copy()) -> list[JobLogMessage]:
     args = script_conf.args
 
     config_path = args.config_path
@@ -112,7 +113,7 @@ def main(script_conf: ScriptConfig = DEFAULT_SCRIPT_CONFIG.copy()) -> list[dict[
     log_directory = config.get('Directory-Section', 'logs')
     ietf_rfc_url = config.get('Web-Section', 'ietf-RFC-tar-private-url')
     temp_dir = config.get('Directory-Section', 'temp')
-    logger = log.get_logger('draftPullLocal', f'{log_directory}/jobs/draft-pull-local.log')
+    logger = log.get_logger('pull_local', f'{log_directory}/jobs/pull-local.log')
     logger.info('Starting cron job IETF pull request local')
 
     messages = []
@@ -120,12 +121,16 @@ def main(script_conf: ScriptConfig = DEFAULT_SCRIPT_CONFIG.copy()) -> list[dict[
     success = True
     try:
         # Clone YangModels/yang repository
-        clone_dir = os.path.join(temp_dir, 'draftpulllocal')
+        clone_dir = os.path.join(temp_dir, 'pull_local')
         if os.path.exists(clone_dir):
             shutil.rmtree(clone_dir)
         repo = repoutil.ModifiableRepoUtil(
-            os.path.join(github_url, 'YangModels/yang.git'),
-            clone_options={'config_username': config_name, 'config_user_email': config_email, 'local_dir': clone_dir},
+            f'{github_url}/YangModels/yang.git',
+            clone_options=repoutil.RepoUtil.CloneOptions(
+                config_username=config_name,
+                config_user_email=config_email,
+                local_dir=clone_dir,
+            ),
         )
         logger.info(f'YangModels/yang repo cloned to local directory {repo.local_dir}')
 
@@ -138,8 +143,7 @@ def main(script_conf: ScriptConfig = DEFAULT_SCRIPT_CONFIG.copy()) -> list[dict[
 
         if tar_opened:
             # Standard RFC modules
-            rfc_path = os.path.join(repo.local_dir, 'standard/ietf/RFC')
-            directory_success, message = populate_directory(rfc_path, notify_indexing, logger)
+            directory_success, message = populate_directory(extract_to, notify_indexing, logger)
             success = success and directory_success
             messages.append({'label': 'Standard RFC modules', 'message': message})
 
@@ -159,7 +163,7 @@ def main(script_conf: ScriptConfig = DEFAULT_SCRIPT_CONFIG.copy()) -> list[dict[
             messages.append({'label': 'IANA modules', 'message': message})
 
     except Exception as e:
-        logger.exception('Exception found while running draftPullLocal script')
+        logger.exception('Exception found while running pull_local script')
         raise e
     if success:
         logger.info('Job finished successfully')