+++ parseAndPopulate/integrity.py
@@ -70,13 +70,6 @@ def check_revision(parsed_module: Statement) -> bool:
     try:
         date(*revision_parts)
     except ValueError:
-        if revision_parts[1:] == [2, 29]:
-            revision_parts[2] = 28
-            try:
-                date(*revision_parts)
-                return True
-            except ValueError:
-                return False
         return False
     return True